requirements.txt
fastapi
uvicorn
openai
python-decouple
